pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "ibis-stream"
version = "0.1.0"
description = "Streaming insert/upsert helpers for Ibis backends"
license = { file = "LICENSE" }
classifiers = [
  "License :: OSI Approved :: Apache Software License",
]
requires-python = ">=3.11,<3.15"
dependencies = [
  "ibis-framework>=12.0.0",
]

[project.optional-dependencies]
bigquery = [
  "google-cloud-bigquery>=3.6.0",
  "google-cloud-bigquery-storage>=2.27.0",
  "ibis-framework[bigquery]>=12.0.0",
]

[dependency-groups]
dev = [
  "pandas>=2.2.0,<3",
  "polars>=1.38.1,<2",
  "pytest>=9.0.2,<10",
  "pytest-asyncio>=1.3.0,<2",
  "ruff>=0.15.2,<0.16",
]

[tool.hatch.build.targets.wheel]
packages = ["src/ibis_stream"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
pythonpath = ["src"]
asyncio_mode = "auto"
markers = [
  "unit: unit tests",
  "integration: tests that require a live instance",
  "bigquery: BigQuery dialect tests",
]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["linux-64", "osx-64", "osx-arm64", "win-64"]

[tool.pixi.dependencies]
python = "3.11.*"

[tool.pixi.environments]
default = ["dev", "bigquery"]

[tool.pixi.tasks]
format = "ruff format ."
format-check = "ruff format --check ."
lint = "ruff check ."
lint-fix = "ruff check --fix ."
test-unit = "pytest -q tests/unit"
test-unit-bigquery = "pytest -q tests/unit/bigquery -m 'unit and bigquery'"
test-integration = "python scripts/integration.py run --dialect all"
test-integration-bigquery = "python scripts/integration.py run --dialect bigquery"
test-bigquery = { depends-on = ["test-unit-bigquery", "test-integration-bigquery"] }
test = { depends-on = ["test-unit", "test-integration"] }
